requirements.txt
streamlit
pandas
requests
plotly
transformers
torch
datasets
python-dotenv
openai
folium
streamlit-folium
huggingface-hub
pydantic
